requirements.txt
opencv_python>=4.5.1.48
Pillow>=9.0.0
numpy>=1.16.0
pip>=21.2.1